@@ -3,7 +3,7 @@ description: A Flutter project demonstrating code size issues with packages.
 publish_to: none
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.2.0
 
 dependencies:
   flutter:
