@@ -4,7 +4,7 @@ description: A Flutter example that uses a platform channel.
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.2.0
 
 dependencies:
   flutter:
