@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 1.0.0+2
 
 environment:
-  sdk: '>=3.0.0'
+  sdk: ^3.2.0
   flutter: '>=3.0.0'
 
 dependencies:
